new: Implement optimized vertex writing
This system replaces the format-specific interfaces that had to
be implemented on a variety of VertexConsumer implementations.

Major speed ups are accomplished by hoisting as much validation
as possible out of the actual vertex writing steps. This allows
the CPU to streamline the process of writing data into memory
and significantly reduces the instruction count.

To write vertices, a specialized "sink" is now created on top
of a "drain" (such a VertexConsumer or BufferBuilder). When
trying to create a sink, the validation logic runs once to
make sure that it's valid to write vertices of the given format
into the drain.

Before writing vertices to a sink, the caller must first make a
call to ensure that there is necessary space in the sink for
the given number of vertices. This will make calls to resize
internal buffers if necessary to fit incoming data.

Once all vertices have been written to the sink, a final call
must be made to flush the vertices in the sink down the drain. If
the caller fails to flush, then the backing storage type will not
be notified that it needs to advance its write pointer or vertex
count, resulting in any previously written vertices being
overwritten. This behavior also enables "safe" zero-copy writing
as a failed write simply gets overwritten by the next caller safely.