Calculate correct lightmap data for blocks with emissive lighting.
Fixes #184. Updates LightDataAccess and FlatLightPipeline to stop ignoring whether or not a block has emissive lighting (currently only MAGMA_BLOCK).  This was caused by the full and opaque condition including emissive and full/opaque blocks, like magma. This resulted in the block being rendered as any other luminant block, and not at full brightness. In addition, the extra condition in FlatLightPipeline#calculate is to prevent the emissive block's faces from using other, incorrect, light data. This is not the cleanest check.