change: Tag build artifacts more thoroughly
The mod's version string now includes extra information about the sources used
to build the mod. If available, it will include the checked out revision in Git
and whether or not the source tree was dirty. If no information about the build
can be determined, the identifier "unknown" will be used.

This is being done because too many players, against our recommendation, are
downloading CI artifacts or compiling the mod themselves for daily use, and
are unable to provide any information about where the build came from when they
inevitably ask for technical support. Not having this information available to us
makes it impossible to diagnose issues over time and creates significant burden.

In the future, this information will be presented in the F3 screen.