fix: Buffer overflow in `UnsafeChunkDrawCallVector`
The `writeEnd` variable used to be set relative to the write pointer, not relative start of the buffer.
This resulted in a `writeEnd` value which was far too large, allowing the buffer to overflow beyond its allocated space
and into random memory, corrupting everything in its way with all the unpredictable doom one would expect from that.