fix: Propagate exceptions from chunk build tasks to the main thread
Chunk build tasks in Sodium can throw due to issues with mods or bugs in chunk
building, so exceptions from these cause the chunk task threads to exit. This
causes the main thread to hang in ChunkRenderManager.updateChunks waiting for
completed build tasks, without creating a crash report for the user to submit.

This change adds a catch-all exception handler around chunk build task
execution, which propagates the exception so that it is re-thrown in
FutureDequeDrain.next() and generates a crash report on the main thread.