fix: Don't use potentially outdated occlusion data for chunk rebuilds
Previously, Sodium would attempt to avoid rebuilding chunks if the occlusion
culler determined that they were not visible. However, when it is time to
rebuild a chunk, the occlusion data is out of date.

Occlusion data is only updated through chunk rebuilds, meaning that if a chunk
is never scheduled for a rebuild, it will never become visible according to
the culler. For an optimization like this to work, the computation of occlusion
data would need to be separated from chunk rebuilds in some way, or the rebuild
logic would need to be smart enough to know when to trust the existing data
and when to ignore it.

For now, it is best to remove this problematic optimization for the sake of
correctness, and once there has been a chance to implement the optimization
in a correct way, it can be restored.

Though the commit de545a500a31f96b5caf7f698a38902945e70f2e makes this issue far
more severe, it is still present in earlier commits. However, for whatever
reason, affected chunks are eventually rebuilt in these previous versions, so
they still pop in, just with some noticeable delay.

Fixes #605 according to my own testing, but testing from others is appreciated
as well.