new: Framework to provide an API to third party mods
The API is separated into its own source set and we build a dedicated jar from that, so mods can depend on it without
depending on the entirety of Sodium (less to download/remap and no accidentally using the implementation classes):
```
dependencies {
    def sodiumVersion = "0.2.0+rev.972c7c1-dirty"
    modCompileOnly "me.jellysquid.mods:sodium-api:${sodiumVersion}"
    if (project.ext.use_sodium_in_dev) {
        modRuntime "me.jellysquid.mods:sodium:${sodiumVersion}"
    }
}
```
For the api jar to be properly remapped, we need to include a dummy fabric.mod.json. Mods should not include the API jar
at runtime (would be weird to have a sodium-api mod present without sodium, and picking the most recent one via the
version will not work for snapshot versions), so that json really is only used for loom.

Not sure what the best way to implement `SodiumApi.get()` is because the api source set cannot reference impl classes
and interfaces cannot contain mutable fields. The best I could come up with is having a semi-internal inner Impl class
with the mutable field.
The API implementation classes are generally (more noticeable in the following commit) separated from other
implementation classes because they - much like the Mixin classes - represent an interface to the outer world, and need
to notify the thingl RenderDevice about entering/leaveing the managed code, i.e. should not be called from other
internal code.

Bumps fabric-loom to 0.7 because 0.6 does not support the Gradle 7 + maven-publish combination (see
https://github.com/FabricMC/fabric-loom/commit/98731532d57d31e3084ef932cce2435d872772c2). Not going straight to 0.8
because that requires Java 16 to run.