new: Simple per-chunk draw parameters for chunk rendering
This technique assigns an index to every chunk in a render region,
and then attaches to every vertex the chunk's index. Using this,
we can index a uniform buffer object to get per-chunk draw parameters.

In effect, this allows us to shave four bytes from the vertex format,
and makes things a little more scalable. Plus, it only depends on GL 3.x
features and is supported everywhere (unlike indirect rendering.)