fix: Prevent rendering sky when fog distance is reduced (#710)
* fix: Prevent rendering sky when fog distance is reduced
When using fog occlusion, it is possible for the sky to be visible through culled chunks, ruining the illusion that there is fog. This is because vanilla relies on chunk fog to hide the sky when fog distance is reduced, however Sodium's chunk occlusion will cull these chunks as they are completely obscured by fog.
To solve this issue without completely disabling chunk occlusion, the sky no longer renders when the fog distance is reduced from the default, since more dense fog means you cannot see the sky past it. Without the sky being rendered, the background of the world becomes the exact same as the fog color, so chunks culled by fog occlusion will now be indistinguishable from the fog. This behavior deviates from vanilla and can cause visual differences, however it solves some visual inconsistencies and matches Minecraft Bedrock Edition (it also seems to hide the sky when the fog distance is reduced). The most notible visual differences from vanilla incude the sky not being visible from underwater and the End dimension's noise sky not being visible during the boss fight, both due to the fog being reduced.
This fixes #35, #438, and #612. It also fixes MC-152504.

* fix: Use LegacyFogHelper instead of GlFogHelper

* change: Update to 1.17

* change: Bump Fabric Loader to 0.11.6

* change: Set source and target compatibility to Java 16

* fix: Use camera status instead of fog distance to detect reduced fog
Before, the fog distance was used to detect if the fog was reduced. While this seems like a good solution, the fog distance is calculated after the sky is rendered, meaning the visibility of the sky was always 1 frame behind. Instead, we can use the camera to calculate if the fog distance is reduced, so the visibility of the sky is up-to-date.

* Rewrite portion of comment describing the patch

I accidentally wrote the exact opposite of the intended message. This commit fixes that, and makes the comment easier to understand by removing unneeded information.

Co-authored-by: JellySquid <jellysquid3@users.noreply.github.com>