fix: Allow chunk rendering mixins to be disabled
This isn't really a complete solution as other mixins will simply be changed
to no-ops without the presence of the chunk renderer, but it at least doesn't
crash. Ideally, the mixin configuration system would be able to see the
dependency chain and simply just not enable dependent mixins.

Note that this configuration isn't really supported, and we're only
providing it for debugging’s sake.

Fixes #842