Added a "Vanilla" tab to the settings panel.
Maybe it would be better served in the bottom corner or in the advanced tab, but considering other mods like Iris add their custom tabs to the top row, I thought it should be done like this for the sake of consistency. In addition, this also allows developers to add their own OptionPageButton via mixin.

I'll write further in my PR.