change: Split OpenGL abstraction layer into different projects
This used to go by the moniker "ThinGL" but it's not so
appropriate anymore.

There is currently no way to supply a different
rendering backend, as Sodium always wants to initialize
the OpenGL backend.

Immediate-mode rendering optimizations have been removed
for the time being, but will be seen again... eventually.

This also gets rid of standalone uniform types, instead
requiring everything to be provided via a UBO.

We still need to deal with std140 layouts in a sane way.