change: Stateless occlusion culling
This allows multiple occlusion culling passes to be
performed in a single frame without making a mess
of global state.

ChunkTree also maintains a tree structure of the
loaded chunks for iterative frustum culling. If
a leaf in the tree is fully visible or not, all
nodes beneath it can be marked the same.