new: Cloud rendering optimizations
Reduces CPU and GPU overhead significantly. Fancy clouds are now 3% to 10% faster
depending on the graphics card, and micro-stutter when moving around should be
slightly better.

Compared to vanilla's Fast clouds, the new Fancy clouds are always faster. As a result,
we are removing the Fast clouds option to reduce maintenance burden.

This also fixes many graphical issues with the vanilla cloud rendering, notably texture
stitches and culling being incorrect.

Fog shading is not yet working for the new cloud rendering, since we need a custom
shader to unlock it.