Add optimized `Direction.getFacing` implementation (#1587)
The old implementation looped over all directions and computed the dot product for each one, looking for the largest. We take advantage of the fact that we're just looking for the closest _basis_ vector (_not_ any arbitrary vector) and compare the input vector's components directly, taking care to preserve the order of precedence for the zero vector.