New driver probe and workarounds for NVIDIA drivers
The NVIDIA drivers now check the command line arguments
for the presence of a sub-string containing ".main", and
forcefully enables their "Threaded Optimizations" hack which
destroys performance and causes crashes with Sodium.

Unfortunately, NVIDIA has implemented this in the most
insane way possible. Rather than using application-specific
profiles, they have instead hardcoded the detection
logic into their driver, and have done so in a manner
that makes it impossible to use NVAPI to override it.

So, in summary, this workaround does two things:

a) We invoke GetCommandLineW and directly modify the pointer
of string data it gives us, in order to strip the command line
arguments of anything identifying.

b) We invoke SetEnvironmentVariableW to set SHIM_MCCOMPAT=0x800000001,
which hints to the NVIDIA driver that it should use the
high-performance GPU. This is required, because as mentioned,
we are breaking the driver's ability to detect Minecraft, which
means it won't try to run it on the high-performance GPU any longer.

Both of these modifications are applied just before context
creation, and any changes are restored right after it. This ensures
other applications which are trying to detect Minecraft are not
affected.

It is worth noting that this problem *only* affected the Vanilla
launcher, since the Fabric Installer chose to include a reference
to the Minecraft entry point on purpose so that the NVIDIA drivers
could continue detecting it. As far as we know, Prism Launcher does
not do this, hence it is not affected by the issue.

As a side effect of this work, we have also improved the
graphics adapter detection code to use OSHI. This is much simpler,
and allows us to enumerate all graphics adapters more reliably. It
also doesn't require a sub-process to be launched, which should
hopefully reduce startup times and some jank.

We have also implemented some minor optimizations which avoid
the calls into glGetInteger/glGetUniformLocation, as we noticed
these were particularly hot in profiling on NVIDIA GPUs.

See #1813 for more information.