Check lightmap when orienting quad
Before, only the brightness (AO) values were checked. Now, if the diagonal brightness value sums are equal, the lightmaps are also checked in a similar way. This fixes the diamond light shape in cases where AO is not present, such as light sources on flat ground and emissive blocks like magma blocks and active sculk sensors.