Fix incorrect chunk tracking with multiple updates in a single frame
When the same chunk is added and removed within the same frame (before the
`loadQueue` could be drained), it could end up in both queues at the same time.
This commit fixes this issue such that any combination of any number of
loads/unloads within a single frame will produce correct results.