Disallow accessing RenderSystem from off-thread
This fixes a lot of potential issues and makes it easier to detect poorly
behaved code. However, this will break certain mods which rely on being
able to call RenderSystem from off-thread.

In my opinion, this is perfectly acceptable breakage to have, as scheduling
individual OpenGL API calls into a shared queue is not safe. If multiple
threads are used, this can result in graphics driver crashes and other
corruption, as the insertion order of commands between threads is
not predictable and cannot possibly be ensured.

If your mod *really* depends on being able to do this, then you should
create a single task with all your graphics commands (rather than a bunch of
individual tasks for every OpenGL command), and submit this to the main
render thread's executor.