Optimize the internals of BufferBuilder
This should be compatible with all normal mods which do not hack
into BufferBuilder or BufferVertexConsumer.

Note: The behavior of the VertexConsumer interface is changed slightly with
these patches, as elements can now be written in any order. However, this
has no consequences, and will not change the behavior of any mod.