Cull entity model faces early when back-facing
We already have the transformed normals and positions for
every cuboid face, so it's possible to use a dot product against the
face normal and any vertex to determine whether it's back-facing.

Since the bottleneck in entity rendering is *primarily* the
number of primitives emitted and uploaded to the GPU, this can
significantly improve frame rates when many entities are visible.