Separate light and face normals to fix some problems
Face normals are the "correct" normals, and are always calculated
using the inverse transpose of the position matrix.

Light normals are what the game typically deals with, and they
are calculated using the normal matrix.

Any code wanting to know the true surface normal of a primitive
should use the face normal, not the light normal.