Implement fingerprinting for donation nags
Fingerprints consist of the Mojang UUID and installation path,
hashed using a securely generated salt. The fingerprint which is
saved to disk cannot be used to recover the data that was used
to create it.

When the fingerprint has been detected as having changed, certain
options will be reset again so that donation nags are shown once more.

This solves problems where (most) mod packs accidentally disable
the config options to show the user our donation nags.