Change handling of version strings
We no longer use the Git repository to tag the builds,
since the metadata isn't useful or reliably present.

Release builds (performed by the CI/CD) always use the
version string as specified in `gradle.properties`.

For non-release builds, the pre-release version will
always be stripped (i.e. `0.6.0-alpha.1` becomes
`0.6.0-snapshot`).

Additionally, the source of the build is tagged at
the end of the version string, either `-local` (for
builds on a local machine) or `-build.N` (for builds
performed by the CI/CD.)