Create it_it
I have taken the pleasure of translating your mod into Italian. Before sending it to you, I checked the consistency with each string in the game menu.

Unfortunately a couple of the strings are long and cover the text area, but you can't do any better!

Thank you for your work!