Improve the terrain vertex format significantly
This solves the following issues:

* Texture coordinates now have a full 15 bits of precision (enough to
represent every texel on a 32k atlas), and the epsilon used to avoid
texture bleeding is added in the shader (rather than being stored in
the quantized value.)

* Vertex positions now have 20-bit precision, fixing the z-fighting
with most resource packs that include block overlays and other very
precise geometry.

* Light values now use a smaller data type (8 bits). This shouldn't
cause issues since it still provides a lot of sub-texel precision.