Additional optimizations for cloud rendering
Since the geometry is always sorted in front-to-back order
now, the Z pre-pass can be eliminated. This may provide a
small performance improvement when fragment bound on the GPU.

Furthermore, we can take advantage of the relative position
of each cloud cell to eliminate back-facing geometry, halving
the total geometry set on average.

The cloud cells near to the camera will also now always have
their interior geometry be rendered, matching the appearance
of Minecraft exactly.