Blacklist ASUS GPU Tweak III from being used
Unfortunately, even with their updates to improve compatibility
with Minecraft, their on-screen overlay *still* uses legacy
OpenGL functions and ends up corrupting graphics state. It does
not work even in Minecraft without modifications, on either AMD
or NVIDIA GPUs.

ASUS, unfortunately, does not provide any documentation on how
application developers can prevent the module from being loaded
into their process. And disabling the On-Screen Overlay (OSD) in
the application settings does not prevent issues.

After reverse engineering their application, it seems the crux
of the problem is their overlay is rendered with legacy OpenGL
functions, which do not work in a Core profile. This is very
similar to the issue with RivaTuner Statistics Server, except
they went a step further to replace the OpenGL context with one
that used a Compatibility profile. We ask *kindly* that ASUS
does not try to do that, and instead fixes their application
to use supported functions in Core profiles.

Until this is resolved, we will blacklist the ASUS GPU Tweak III
application entirely.